refactor - restukturyzacja projektu na oddizlene pliki i foldery
