+++ include/chess/rules/Attack.h
@@ -1,0 +1,4 @@
+// Attack.h
+#pragma once
+#include "chess/board/Board.h"
+
