+++ include/chess/rules/Castling.h
@@ -1,0 +1,6 @@
+// Castling.h
+#pragma once
+#include <string>
+#include "chess/board/Board.h"
+#include "chess/model/Move.h"
+
