+++ include/chess/rules/MoveExec.h
@@ -1,0 +1,4 @@
+// MoveExec.h
+#pragma once
+#include "chess/board/Board.h"
+#include "chess/model/Move.h"
