+++ include/chess/rules/MoveValid.h
@@ -1,0 +1,5 @@
+// MoveValid.h
+#pragma once
+#include "chess/board/Board.h"
+#include "chess/model/Move.h"
+
